+++ test/Utils/DetailFunctions.php
@@ -1,0 +1,14 @@
+<?php
+namespace Test\Utils;
+
+use TRegx\CleanRegex\Match\Details\Detail;
+
+class DetailFunctions
+{
+    public static function text(): callable
+    {
+        return static function (Detail $detail): string {
+            return $detail->text();
+        };
+    }
+}